requirements.txt
python-jenkins
requests
